# docs/swagger.yaml
basePath: /
definitions:
...
  server.swaggerAppointmentResponse:
    properties:
      id:
        type: string
    type: object
  server.swaggerAuditTrailResponse:
    properties:
      action:
        type: string
      actor_id:
        type: string
      actor_name:
        type: string
      actor_role:
        type: string
      appointment_id:
        type: string
      changes:
        type: string
      created_at:
        type: string
      entity_id:
        type: string
      entity_type:
        type: string
      id:
        type: string
      message:
        type: string
    type: object
  server.swaggerCreateAppointmentRequest:
    properties:
      description:
        type: string
      doctor_id:
        type: string
      hospital_id:
        type: string
      patient_id:
        type: string
      timeslot_id:
        type: string
    type: object
...
  server.swaggerNoteResponse:
    properties:
      appointment_id:
        type: string
      content:
        type: string
      created_at:
        type: string
      id:
        type: string
      updated_at:
        type: string
    type: object
  server.swaggerPrescriptionResponse:
    properties:
      appointment_id:
        type: string
      details:
        type: string
      doctor_id:
        type: string
      expiration_date:
        type: string
      hospital_id:
        type: string
      id:
        type: string
      patient_id:
        type: string
      status:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@clinicwise.io
    name: Clinic Wise Support
  description: REST API for the Clinic Wise healthcare management platform.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Clinic Wise API
  version: "1.0"
paths:
  /appointments:
    get:
      description: Returns all appointments for the given hospital. Requires authentication.
      parameters:
      - description: Hospital ID (ULID)
        in: query
        name: hospital_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/server.swaggerAppointmentResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List appointments for a hospital
      tags:
      - Appointments
    post:
      consumes:
      - application/json
      description: Creates a new appointment. Requires authentication.
      parameters:
      - description: Create appointment payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/server.swaggerCreateAppointmentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.swaggerAppointmentResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create an appointment
      tags:
      - Appointments
  /appointments/{id}:
    get:
      description: Returns a single appointment by ID. Requires authentication.
      parameters:
      - description: Appointment ID (ULID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.swaggerAppointmentResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get an appointment
      tags:
      - Appointments
  /appointments/{id}/complete:
    patch:
      description: Marks an appointment as completed. Requires doctor role.
      parameters:
      - description: Appointment ID (ULID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.swaggerAppointmentResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Complete an appointment
      tags:
      - Appointments
  /appointments/user:
    get:
      description: Returns paginated appointments belonging to the authenticated user.
      parameters:
      - description: Page number (1-based)
        in: query
        name: page
        required: true
        type: integer
      - description: Page size
        in: query
        name: limit
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/server.swaggerAppointmentResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List appointments for the current user
      tags:
      - Appointments
  /audit-trails/appointment/{id}:
    get:
      description: Returns audit log entries for a given appointment. Requires authentication
        and ownership.
      parameters:
      - description: Appointment ID (ULID)
        in: path
        name: id
        required: true
        type: string
      - description: Filter by action
        in: query
        name: action
        type: string
      - description: Filter by actor ID (ULID)
        in: query
        name: actor_id
        type: string
      - description: Filter from date (RFC3339)
        in: query
        name: from
        type: string
      - description: Filter to date (RFC3339)
        in: query
        name: to
        type: string
      - description: Page number (1-based)
        in: query
        name: page
        type: integer
      - description: Page size (max 200, default 50)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/server.swaggerAuditTrailResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List audit trail for an appointment
      tags:
      - Audit Trails
  /audit-trails/entity/{type}/{id}:
    get:
      description: Returns audit log entries for a given entity type and ID. Requires
        authentication.
      parameters:
      - description: Entity type (e.g. user, appointment)
        in: path
        name: type
        required: true
        type: string
      - description: Entity ID
        in: path
        name: id
        required: true
        type: string
      - description: Filter by action
        in: query
        name: action
        type: string
      - description: Filter by actor ID (ULID)
        in: query
        name: actor_id
        type: string
      - description: Filter from date (RFC3339)
        in: query
        name: from
        type: string
      - description: Filter to date (RFC3339)
        in: query
        name: to
        type: string
      - description: Page number (1-based)
        in: query
        name: page
        type: integer
      - description: Page size (max 200, default 50)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/server.swaggerAuditTrailResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List audit trail for an entity
      tags:
      - Audit Trails
...
  /notes/appointment/{id}:
    get:
      description: Returns all notes for a given appointment. Requires doctor or patient
        role and ownership.
      parameters:
      - description: Appointment ID (ULID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/server.swaggerNoteResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List notes for an appointment
      tags:
      - Notes
...
  /prescriptions/appointment/{id}:
    get:
      description: Returns all prescriptions for a given appointment. Requires doctor
        or patient role.
      parameters:
      - description: Appointment ID (ULID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/server.swaggerPrescriptionResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List prescriptions for an appointment
      tags:
      - Prescriptions
securityDefinitions:
  BearerAuth:
    description: 'JWT Bearer token. Example: "Bearer <token>"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"